[Gold V] Title: KSA 문자열, Time: 4 ms, Memory: 2912 KB, Score: 40 point -BaekjoonHub

+++ 백준/Gold/31433. KSA 문자열/KSA 문자열.cc
@@ -12,7 +12,7 @@
 using namespace std;
 string s;
 string check[3] = {"KSA", "SAK", "AKS"};
-int cnt[3];
+long long cnt[3];
 
 int main()
 {
